2021/2/25

+++ ehis-hems-service/ehis-hems-claimflow/src/main/java/com/paic/ehis/claimflow/mapper/ClaimCasePayeeMapper.java
@@ -4,6 +4,7 @@ package com.paic.ehis.claimflow.mapper;
 
 import com.paic.ehis.claimflow.domain.ClaimCasePayee;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -76,5 +77,5 @@ public interface ClaimCasePayeeMapper
      */
     public List<ClaimCasePayee> selectClaimCasePayeeByRptNo(String rptNo);
 
-    public int CheckClaimCasePayeeByRptNo(String rptNo);
+    public BigDecimal CheckClaimCasePayeeByRptNo(String rptNo);
 }
